Shelley: split off Query in a separate module
The initial reason for doing this was to reduce the atrocious compile time of
the `Ouroboros.Consensus.Shelley.Ledger.Ledger`, 3 min on my machine!

Splitting off `Query` didn't help much, that module now takes 2 min 30 sec to
compile on my machine.

Nevertheless, splitting it off makes sense, even if it doesn't reduce compile
time much.
