+++ ouroboros-consensus-shelley/ouroboros-consensus-shelley.cabal
@@ -36,6 +36,7 @@ library
                        Ouroboros.Consensus.Shelley.Ledger.Ledger
                        Ouroboros.Consensus.Shelley.Ledger.Mempool
                        Ouroboros.Consensus.Shelley.Ledger.NetworkProtocolVersion
+                       Ouroboros.Consensus.Shelley.Ledger.Query
                        Ouroboros.Consensus.Shelley.Ledger.TPraos
                        Ouroboros.Consensus.Shelley.Node
                        Ouroboros.Consensus.Shelley.Node.Serialisation
