+++ ouroboros-consensus-shelley/src/Ouroboros/Consensus/Shelley/Ledger.hs
@@ -9,3 +9,4 @@ import           Ouroboros.Consensus.Shelley.Ledger.Integrity as X
 import           Ouroboros.Consensus.Shelley.Ledger.Ledger as X
 import           Ouroboros.Consensus.Shelley.Ledger.Mempool as X
 import           Ouroboros.Consensus.Shelley.Ledger.NetworkProtocolVersion as X
+import           Ouroboros.Consensus.Shelley.Ledger.Query as X
